Implement enough for CLI tests to pass
Except cli_version, because for some reason clap's `version` magic
attribute prints to stderr, instead of stdout as the test requires. Any
attempts to manually define a version flag cause only the application
name to be printed to stderr. Flags not called `version` work just fine
though.

There might be a way forward with the Builder API, or another attribute
to configure where the version information is directed.